+++ engine/src/components/Controls/Controls.js
@@ -4,6 +4,7 @@ const moveIcon = require('../../../dist/icons/moveIcon.svg');
 const selectIcon = require('../../../dist/icons/selectIcon.svg');
 const boxZoomIcon = require('../../../dist/icons/boxZoomIcon.svg');
 const resetIcon = require('../../../dist/icons/resetIcon.svg');
+//const arrowUpIcon = require('../../../dist/icons/arrowUpIcon.svg');
 
 const Controls = function(config) {
   this.graph = config.graph;
@@ -37,6 +38,20 @@ const Controls = function(config) {
     icon: zoomOutIcon,
     evtName: 'zoom-out'
   });
+
+
+  // if (config.showStepControls) {
+  //   this.stepUpButton = this.addButton({
+  //     icon: arrowUpIcon,
+  //     evtName: 'step-up'
+  //   });
+  //   this.stepDownButton = this.addButton({
+  //     icon: arrowUpIcon,
+  //     evtName: 'step-down'
+  //   });
+
+
+  // }
 
 
 };