+++ webpack.config.js
@@ -1,11 +1,11 @@
 module.exports = {
   entry: './engine/src/ElGrapho.js',
   output: {
-    filename: 'ElGrapho.js',
     path: __dirname + '/engine/dist',
+    filename: 'ElGrapho.js',
     library: 'ElGrapho',
     libraryTarget: 'umd',
-    globalObject: 'typeof window !== \'undefined\' ? window : this'
+    globalObject: 'this'
   },
   devtool: 'source-map',
   mode: 'development' // development or production