+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Changlog
 
-## v1.4.1
+## v2.1.0
+* models are now called layouts
+
+## v2.0.0
+* License is now MIT!
+* new API
+* ForceDirected model now using Webcola for layout
+* new Web model which is a simplified force directed graph that runs in O(n) time
+* node color cycling
+* auto viewport fitting
+
+## v1.6.0
+* new edges API (see docs)
+* labels for nodes
+* edge arrows for directed graphs
+
+## v1.5.4
+* fix npm export
+
+## v1.5.0
+* new ForceDirectedGraph model
 * new Ring model
 * circle nodes
 * node strokes